create page balance

+++ database/migrations/2022_08_20_063430_create_credit_balances_table.php
@@ -16,7 +16,7 @@ class CreateCreditBalancesTable extends Migration
         Schema::create('credit_balances', function (Blueprint $table) {
             $table->id();
             $table->foreignId('users_id')->constrained('users')->onDelete('cascade');
-            $table->bigInteger('amount');
+            $table->decimal('amount', 13, 3);
             $table->timestamps();
         });
     }
